BG P5 (1/n): extract Config.swift domain types into Models/
Config.swift carried eight top-level model types before the `struct Config`
settings container even began. They now live one file each under Sources/Models/
(matching the existing convention there):

  LiveAIMode, QuickAction, GatewayProvider + GatewayConfig, ModelConfig,
  PromptPreset, Persona, CustomToolDefinition, AppMode

These were already top-level types (not nested in Config), so moving them changes
no call sites — same module, same names, same access. Config.swift drops from
3,209 to 2,810 lines; a stray misplaced doc comment was tidied.

Pure code motion, no behaviour change. The riskier P5 work (@UserDefaultsBacked
property wrapper + per-feature settings structs, filing loose Services/ files,
RemindersTool→NotesTool rename) is deferred to follow-up increments.

Suite: 1513 tests, 0 failures; Release green.
