+++ OpenGlasses/Sources/Utils/Config.swift
@@ -1,376 +1,5 @@
 import Foundation
 
-/// A LiveAI mode preset that changes the system instruction for realtime sessions.
-struct LiveAIMode: Codable, Identifiable, Hashable {
-    var id: String
-    var name: String
-    var icon: String
-    var promptPrefix: String
-
-    static let builtIn: [LiveAIMode] = [
-        LiveAIMode(id: "standard", name: "Standard", icon: "bubble.left", promptPrefix: ""),
-        LiveAIMode(id: "museum", name: "Museum Guide", icon: "building.columns", promptPrefix: "You are acting as a museum docent and art expert. When the user shows you artwork, sculptures, or exhibits, identify them and provide engaging context: the artist, period, technique, and cultural significance. Be enthusiastic and educational.\n\n"),
-        LiveAIMode(id: "accessibility", name: "Blind Assistant", icon: "figure.walk", promptPrefix: "You are a visual accessibility assistant for a visually impaired user. Describe the environment in detail: obstacles, people, signage, doors, stairs, vehicles, and spatial layout. Be specific about distances and directions (left, right, ahead). Prioritize safety-critical information.\n\n"),
-        LiveAIMode(id: "reading", name: "Reading Assistant", icon: "text.viewfinder", promptPrefix: "You are a reading assistant. Focus on any visible text — signs, menus, documents, labels, screens. Read text aloud clearly and completely. For foreign languages, first read the original, then translate. Offer to summarize long text.\n\n"),
-        LiveAIMode(id: "translator", name: "Live Translator", icon: "globe", promptPrefix: "You are a real-time translator. When you see text or hear speech in a foreign language, translate it naturally. Provide the original text first, then the translation. For signs and menus, translate everything visible.\n\n"),
-        LiveAIMode(id: "tutor", name: "Language Tutor", icon: "graduationcap", promptPrefix: "You are a language tutor. Help the user practice the language of the text/signs they show you. Pronounce words clearly, explain grammar, suggest phrases for the situation. Be encouraging and patient.\n\n"),
-        LiveAIMode(id: "golf", name: "Golf Caddy", icon: "figure.golf", promptPrefix: "You are a golf caddy on smart glasses. Help with club selection, read greens, track shots, and provide course strategy. Be confident and decisive. Keep advice brief during play — 1-2 sentences per decision. Celebrate good shots, stay positive on bad ones.\n\n"),
-    ]
-}
-
-/// A user-configurable quick action button shown on the main screen.
-struct QuickAction: Codable, Identifiable {
-    var id: String
-    var label: String
-    var icon: String
-    var type: ActionType
-
-    enum ActionType: String, Codable, CaseIterable, Identifiable {
-        case prompt = "prompt"
-        case photo = "photo"
-        case photoThenPrompt = "photoThenPrompt"
-        case homeAssistant = "homeAssistant"
-        case siriShortcut = "siriShortcut"
-        case openApp = "openApp"
-
-        var id: String { rawValue }
-
-        var displayName: String {
-            switch self {
-            case .prompt: return "Text Prompt"
-            case .photo: return "Take Photo"
-            case .photoThenPrompt: return "Photo + Prompt"
-            case .homeAssistant: return "Home Assistant"
-            case .siriShortcut: return "Siri Shortcut"
-            case .openApp: return "Open App"
-            }
-        }
-
-        var description: String {
-            switch self {
-            case .prompt: return "Send a text prompt to the AI"
-            case .photo: return "Capture and describe a photo"
-            case .photoThenPrompt: return "Capture a photo with a custom prompt"
-            case .homeAssistant: return "Call a Home Assistant service"
-            case .siriShortcut: return "Run a Siri Shortcut by name"
-            case .openApp: return "Open an app via URL scheme"
-            }
-        }
-    }
-
-    /// The prompt text (for .prompt and .photoThenPrompt types)
-    var promptText: String?
-    /// Home Assistant service call (e.g., "light.turn_off") for .homeAssistant type
-    var haService: String?
-    /// Home Assistant entity ID (e.g., "light.living_room") for .homeAssistant type
-    var haEntityId: String?
-    /// Extra data as JSON string for .homeAssistant type (e.g., {"brightness": 50})
-    var haData: String?
-    /// Siri Shortcut name for .siriShortcut type
-    var shortcutName: String?
-    /// URL scheme for .openApp type (e.g., "weixin://")
-    var urlScheme: String?
-
-    static let travelTemplates: [QuickAction] = [
-        QuickAction(
-            id: "travel-translate-sign-menu",
-            label: "Translate Sign",
-            icon: "text.viewfinder",
-            type: .photoThenPrompt,
-            promptText: "Read all visible text in this image. First provide exact original text, then translate to English. If helpful, use the translate tool to improve accuracy. Keep response concise for glasses."
-        ),
-        QuickAction(
-            id: "travel-ask-local-phrase",
-            label: "Local Phrase",
-            icon: "globe",
-            type: .prompt,
-            promptText: "Help me say this naturally in the local language where I am. If my intent is unclear, ask one short clarification first. Then provide local phrase, pronunciation, and a polite variant. Use the translate tool."
-        ),
-    ]
-
-    /// Built-in Field Assist quick action. Injected at the front of `Config.quickActions`
-    /// whenever Field Assist is active (see `withFieldAssistAction`) — it is never persisted,
-    /// so it appears/disappears with the entitlement. A `.prompt` action so it routes through
-    /// the existing pipeline and the AI starts the session via the `field_session` tool.
-    static let fieldAssist = QuickAction(
-        id: "field-assist",
-        label: "Field Assist",
-        icon: "wrench.and.screwdriver.fill",
-        type: .prompt,
-        promptText: "Start a Field Assist session on my default vault. Briefly confirm you're ready and what you can help me troubleshoot."
-    )
-
-    static let defaults: [QuickAction] = [
-        QuickAction(id: "describe", label: "Describe", icon: "eye", type: .photoThenPrompt,
-                    promptText: "Describe what you see in this image in detail."),
-        QuickAction(id: "calendar", label: "Event", icon: "calendar", type: .photoThenPrompt,
-                    promptText: "Extract any event details from this image (dates, times, locations, names) and create a calendar entry summary."),
-        QuickAction(id: "task", label: "Task", icon: "checklist", type: .photoThenPrompt,
-                    promptText: "Extract any action items or tasks from this image and list them."),
-        QuickAction(id: "lights-off", label: "Lights Off", icon: "lightbulb.slash", type: .homeAssistant,
-                    haService: "light.turn_off", haEntityId: "all"),
-    ] + travelTemplates
-}
-
-/// A saved LLM model configuration
-// MARK: - Gateway Configuration
-
-/// Known gateway providers — users can also add custom ones.
-enum GatewayProvider: String, Codable, CaseIterable, Identifiable {
-    case openclaw = "openclaw"
-    case nanoclaw = "nanoclaw"
-    case nemoclaw = "nemoclaw"
-    case custom = "custom"
-
-    var id: String { rawValue }
-
-    var displayName: String {
-        switch self {
-        case .openclaw: return "OpenClaw"
-        case .nanoclaw: return "NanoClaw"
-        case .nemoclaw: return "NemoClaw"
-        case .custom: return "Custom"
-        }
-    }
-
-    var icon: String {
-        switch self {
-        case .openclaw: return "server.rack"
-        case .nanoclaw: return "desktopcomputer"
-        case .nemoclaw: return "cpu"
-        case .custom: return "gear"
-        }
-    }
-
-    var defaultPort: Int {
-        switch self {
-        case .openclaw: return 18789
-        case .nanoclaw: return 18789
-        case .nemoclaw: return 18789
-        case .custom: return 18789
-        }
-    }
-
-    /// Whether this provider uses the standard OpenClaw WebSocket protocol.
-    var usesOpenClawProtocol: Bool { true }
-}
-
-/// A configured gateway endpoint — could be OpenClaw, NanoClaw, NemoClaw, or custom.
-struct GatewayConfig: Codable, Identifiable, Equatable {
-    var id: String
-    var name: String                    // User label, e.g. "Mac Mini OpenClaw"
-    var provider: String                // GatewayProvider rawValue
-    var lanHost: String                 // LAN/local IP or hostname
-    var port: Int                       // Default 18789
-    var tunnelHost: String              // Tailscale or tunnel URL
-    var token: String                   // Pre-shared gateway auth token (legacy/manual path)
-    var connectionMode: String          // "auto", "lan", "tunnel"
-    var enabled: Bool
-    var priority: Int                   // Lower = tried first
-    // Device pairing (optional w/ nil default → backward-compatible with previously-saved
-    // gateways and with every existing memberwise-init call site):
-    var deviceToken: String? = nil      // Per-device token issued by the gateway after approval
-    var deviceId: String? = nil         // Stable per-device identity sent in the handshake
-    var setupCode: String? = nil        // Transient bootstrap code, cleared once a device token lands
-
-    var gatewayProvider: GatewayProvider {
-        GatewayProvider(rawValue: provider) ?? .custom
-    }
-
-    var connectionModeEnum: OpenClawConnectionMode {
-        OpenClawConnectionMode(rawValue: connectionMode) ?? .auto
-    }
-
-    /// Build the LAN URL from host + port.
-    var lanURL: String {
-        let host = lanHost.trimmingCharacters(in: CharacterSet(charactersIn: "/"))
-        guard !host.isEmpty else { return "" }
-        return "\(host):\(port)"
-    }
-
-    /// The tunnel URL (already includes port usually).
-    var tunnelURL: String {
-        tunnelHost.trimmingCharacters(in: CharacterSet(charactersIn: "/"))
-    }
-
-    /// True when there's a usable credential (shared token *or* a paired device token) and a
-    /// host to reach. A device-paired gateway needs no shared token.
-    var isConfigured: Bool {
-        let hasCredential = !token.isEmpty || !(deviceToken ?? "").isEmpty
-        return hasCredential && (!lanHost.isEmpty || !tunnelHost.isEmpty)
-    }
-
-    /// Create a new gateway with defaults for a given provider.
-    static func newGateway(provider: GatewayProvider, priority: Int = 0) -> GatewayConfig {
-        GatewayConfig(
-            id: UUID().uuidString,
-            name: "\(provider.displayName) Gateway",
-            provider: provider.rawValue,
-            lanHost: "",
-            port: provider.defaultPort,
-            tunnelHost: "",
-            token: "",
-            connectionMode: "auto",
-            enabled: true,
-            priority: priority
-        )
-    }
-}
-
-struct ModelConfig: Codable, Identifiable, Equatable {
-    var id: String  // UUID string
-    var name: String  // User-facing label, e.g. "Claude Sonnet" or "GPT-4o"
-    var provider: String  // LLMProvider rawValue
-    var apiKey: String
-    var model: String
-    var baseURL: String
-    /// Optional user override for whether this model accepts image input.
-    /// When nil, the app falls back to provider/model-name heuristics.
-    var supportsVision: Bool? = nil
-
-    /// Convenience to get the LLMProvider enum
-    var llmProvider: LLMProvider {
-        LLMProvider(rawValue: provider) ?? .custom
-    }
-
-    /// Whether this model should receive image input when the app has an image available.
-    var visionEnabled: Bool {
-        supportsVision ?? Self.inferredSupportsVision(provider: llmProvider, model: model, baseURL: baseURL)
-    }
-
-    static func inferredSupportsVision(provider: LLMProvider, model: String, baseURL: String) -> Bool {
-        switch provider {
-        case .anthropic, .gemini, .openai:
-            return true
-        case .groq, .local, .appleOnDevice:
-            return false
-        case .qwen:
-            // Qwen3.5-plus and qwen-vl models support vision
-            let lowerModel = model.lowercased()
-            return lowerModel.contains("vl") || lowerModel.contains("plus") || lowerModel.contains("max") || lowerModel.contains("omni")
-        case .xai:
-            // Grok 4 family is multimodal; earlier Grok text models are not
-            let lowerModel = model.lowercased()
-            return lowerModel.contains("grok-4") || lowerModel.contains("vision")
-        case .openrouter:
-            // OpenRouter supports vision for many models
-            let lowerModel = model.lowercased()
-            return lowerModel.contains("claude") || lowerModel.contains("gpt-4") || lowerModel.contains("gemini") || lowerModel.contains("vision") || lowerModel.contains("llava")
-        case .zai, .minimax, .custom:
-            let lowerModel = model.lowercased()
-            let lowerBaseURL = baseURL.lowercased()
-
-            let knownVisionHints = [
-                "vision", "gpt-4", "gpt-4.1", "gpt-4o", "o1", "o3",
-                "claude-3", "claude-4", "sonnet", "opus",
-                "gemini", "vl", "qwen-vl", "qwen2.5-vl", "qvq",
-                "pixtral", "llava", "minicpm-v", "glm-4.1v"
-            ]
-
-            if knownVisionHints.contains(where: { lowerModel.contains($0) }) {
-                return true
-            }
-
-            if lowerBaseURL.contains("openrouter.ai") {
-                return knownVisionHints.contains(where: { lowerModel.contains($0) })
-            }
-
-            return false
-        }
-    }
-
-    /// Create a new config with defaults for a provider
-    static func defaultConfig(for provider: LLMProvider) -> ModelConfig {
-        ModelConfig(
-            id: UUID().uuidString,
-            name: provider.displayName,
-            provider: provider.rawValue,
-            apiKey: "",
-            model: provider.defaultModel,
-            baseURL: provider.defaultBaseURL
-        )
-    }
-}
-
-/// A saved system prompt preset
-struct PromptPreset: Codable, Identifiable, Equatable {
-    var id: String
-    var name: String
-    var prompt: String
-    var isBuiltIn: Bool
-    var icon: String?
-    /// Suggested camera behavior for this preset mode.
-    /// "smart" = auto-activate on vision queries, "always" = keep camera on, nil = default behavior.
-    var cameraBehavior: String?
-}
-
-/// A persona bundles a wake word, AI model, and system prompt.
-/// Multiple personas can be active simultaneously — each wake word routes to its own model+prompt.
-struct Persona: Codable, Identifiable, Equatable {
-    var id: String
-    var name: String                      // "Claude", "Jarvis", "Computer"
-    var wakePhrase: String                // "hey claude"
-    var alternativeWakePhrases: [String]   // ["hey cloud", "hey claud"]
-    var modelId: String                   // References ModelConfig.id
-    var presetId: String                  // References PromptPreset.id
-    var enabled: Bool
-    /// SF Symbol icon name for display in persona picker / mode cards.
-    var icon: String?
-    /// Whether this is a built-in preset persona (shipped with the app).
-    var isBuiltIn: Bool?
-
-    // MARK: - Agentic Capabilities (optional)
-
-    /// Custom soul.md content for this persona. When set, overrides the global soul.
-    var soulOverride: String?
-
-    /// Chattiness level for this persona (nil = use global setting).
-    /// Raw string matching Config.AgentChattiness: "quiet", "normal", "chatty".
-    var chattinessRaw: String?
-
-    /// Specific tools this persona can use (nil = all tools). Restrict to subset for focused agents.
-    var allowedTools: [String]?
-
-    /// Scheduled task IDs this persona owns. Tasks only run when this persona is active.
-    var ownedTaskIds: [String]?
-
-    /// All phrases this persona responds to (primary + alternatives).
-    var allPhrases: [String] {
-        [wakePhrase] + alternativeWakePhrases
-    }
-}
-
-/// A user-defined tool that maps to a Siri Shortcut or URL scheme
-struct CustomToolDefinition: Codable, Identifiable, Equatable {
-    var id: String
-    var name: String
-    var description: String
-    var parameters: [CustomToolParam]
-    var actionType: ActionType
-    var shortcutName: String?
-    var urlTemplate: String?
-
-    enum ActionType: String, Codable, CaseIterable {
-        case shortcut
-        case urlScheme
-
-        var displayName: String {
-            switch self {
-            case .shortcut: return "Siri Shortcut"
-            case .urlScheme: return "URL Scheme"
-            }
-        }
-    }
-
-    struct CustomToolParam: Codable, Identifiable, Equatable {
-        var id: String
-        var name: String
-        var type: String  // "string" or "number"
-        var description: String
-        var required: Bool
-    }
-}
-
 /// App configuration and API keys
 struct Config {
     /// Anthropic API key for Claude. Stored in the Keychain (see `KeychainService`).
@@ -3179,31 +2808,3 @@ struct Config {
     }
 }
 
-// MARK: - App Mode Enum
-
-enum AppMode: String, CaseIterable {
-    case direct = "direct"
-    case geminiLive = "geminiLive"
-    case openaiRealtime = "openaiRealtime"
-
-    var displayName: String {
-        switch self {
-        case .direct: return "Direct Mode"
-        case .geminiLive: return "Gemini Live"
-        case .openaiRealtime: return "OpenAI Realtime"
-        }
-    }
-
-    var description: String {
-        switch self {
-        case .direct: return "Wake word, any LLM provider, text-to-speech"
-        case .geminiLive: return "Real-time audio/video streaming via Gemini"
-        case .openaiRealtime: return "Real-time audio/video streaming via OpenAI"
-        }
-    }
-
-    /// Whether this mode is a real-time streaming mode (as opposed to wake-word direct mode).
-    var isRealtime: Bool {
-        self == .geminiLive || self == .openaiRealtime
-    }
-}
